requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
Flask-Login>=0.5
Flask-WTF>=1.0
python-dotenv>=0.19
bcrypt>=3.2
psycopg2-binary>=2.9.5
requests>=2.28.0
supabase>=1.0.0
gunicorn>=20.1.0
Werkzeug>=2.0.0
Jinja2>=3.0.0
itsdangerous>=2.0.0
click>=8.0.0
MarkupSafe>=2.0.0
pandas>=1.5.0
python-docx>=0.8.11
PyPDF2>=2.10.0
openpyxl>=3.0.10
WTForms>=3.0.0